requirements.txt
joblib==0.17.0
pandas==1.1.4
toml==0.10.2
seaborn==0.11.0
PyGithub==1.54
wrapt==1.12.1
requests==2.24.0
python-dotenv==0.15.0
aiohttp==3.7.3